requirements.txt
fastapi==0.94.1
FastAPI-SQLAlchemy==0.2.1
alembic==1.10.2
pydantic==1.10.6
SQLAlchemy==2.0.6
starlette==0.26.1
uvicorn==0.21.1
PyMySQL==1.0.2
mysql-connector-python==8.0.32
chardet==5.1.0
python-dotenv==1.0.0
email-validator==1.3.1